# api_v2_backup_20251016_203324/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.MovieResponse:
    properties:
      adult:
        type: boolean
      backdrop_url:
        type: string
      external_id:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      overview:
        type: string
      poster_url:
        type: string
      release_date:
        type: string
      runtime:
        type: integer
      title:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.MovieSearchResponse:
    properties:
      movies:
        items:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MovieResponse'
        type: array
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewMovieResponse:
    properties:
      adult:
        type: boolean
      backdrop_url:
        type: string
      external_id:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      overview:
        type: string
      poster_url:
        type: string
      release_date:
        type: string
      runtime:
        type: integer
      title:
        type: string
      vote_average:
        type: number
      vote_count:
        type: integer
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      movie:
        $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewMovieResponse'
      rating:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.UserProfile'
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.UserProfile:
    properties:
      bio:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      is_private:
        type: boolean
      profile_picture_url:
        type: string
      theme:
        type: string
      username:
        type: string
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cineverse.com
    name: CineVerse API Support
    url: http://www.cineverse.com/support
  description: A comprehensive social network platform for movie enthusiasts
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://cineverse.com/terms/
  title: CineVerse API
  version: "2.0"
paths:
...
  /movies/{id}:
    get:
      description: Get movie details by internal database ID
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get movie by ID
      tags:
      - Movies
  /movies/external/{externalId}:
    get:
      description: Get movie details by TMDb external ID
      parameters:
      - description: TMDb Movie ID
        in: path
        name: externalId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get movie by external ID
      tags:
      - Movies
  /movies/genre/{genre}:
    get:
      description: Get movies filtered by genre
      parameters:
      - description: Genre name
        in: path
        name: genre
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MovieSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get movies by genre
      tags:
      - Movies
  /movies/popular:
    get:
      description: Get a list of popular movies
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MovieSearchResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get popular movies
      tags:
      - Movies
  /movies/search:
    get:
      consumes:
      - application/json
      description: Search for movies by title
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MovieSearchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Search movies
      tags:
      - Movies
...
  /reviews/movie/{movieId}:
    get:
      description: Get all reviews for a specific movie with pagination
      parameters:
      - description: Movie ID
        in: path
        name: movieId
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of movie reviews
          schema:
            items:
              $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ReviewResponse'
            type: array
        "400":
          description: Invalid movie ID
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "404":
          description: Movie not found
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Get movie reviews
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"